Refactor UpdateChannels
